+++ enterprise/cmd/frontend/internal/auth/githubappauth/resolver_test.go
@@ -2,6 +2,7 @@ package githubapp
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"testing"
@@ -17,8 +18,10 @@ import (
 	ghtypes "github.com/sourcegraph/sourcegraph/enterprise/internal/github_apps/types"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/types"
+	"github.com/sourcegraph/sourcegraph/schema"
 )
 
 // userCtx returns a context where give user ID identifies logged in user.
@@ -124,12 +127,6 @@ func TestResolver_GitHubApps(t *testing.T) {
 			ID: 1,
 		},
 	}, nil)
-	gitHubAppsStore.DeleteFunc.SetDefaultHook(func(ctx context.Context, app int) error {
-		if app != id {
-			return errors.New(fmt.Sprintf("app with id %d does not exist", id))
-		}
-		return nil
-	})
 
 	db := edb.NewStrictMockEnterpriseDB()
 
@@ -175,6 +172,99 @@ func TestResolver_GitHubApps(t *testing.T) {
 		ExpectedErrors: []*gqlerrors.QueryError{{
 			Message: "must be site admin",
 			Path:    []any{string("gitHubApps")},
+		}},
+	}})
+}
+
+func TestResolver_GitHubApp(t *testing.T) {
+	logger := logtest.Scoped(t)
+	id := 1
+	graphqlID := MarshalGitHubAppID(int64(id))
+
+	userStore := database.NewMockUserStore()
+	userStore.GetByCurrentAuthUserFunc.SetDefaultHook(func(ctx context.Context) (*types.User, error) {
+		a := actor.FromContext(ctx)
+		if a.UID == 1 {
+			return &types.User{ID: 1, SiteAdmin: true}, nil
+		}
+		if a.UID == 2 {
+			return &types.User{ID: 2, SiteAdmin: false}, nil
+		}
+		return nil, errors.New("not found")
+	})
+
+	gitHubAppsStore := store.NewStrictMockGitHubAppsStore()
+	gitHubAppsStore.GetByIDFunc.SetDefaultReturn(&ghtypes.GitHubApp{
+		ID:      1,
+		AppID:   1,
+		BaseURL: "https://github.com",
+	}, nil)
+
+	extsvcStore := database.NewMockExternalServiceStore()
+	conn := &schema.GitHubConnection{
+		Url: "https://github.com",
+		GitHubAppDetails: &schema.GitHubAppDetails{
+			AppID:          1,
+			InstallationID: 123,
+		},
+	}
+	extsvcConn, err := json.Marshal(conn)
+	require.NoError(t, err)
+	extsvcStore.ListFunc.SetDefaultReturn([]*types.ExternalService{{
+		DisplayName: "MockExtsvc",
+		Kind:        extsvc.KindGitHub,
+		Config:      extsvc.NewUnencryptedConfig(string(extsvcConn)),
+	}}, nil)
+
+	db := edb.NewStrictMockEnterpriseDB()
+
+	db.GitHubAppsFunc.SetDefaultReturn(gitHubAppsStore)
+	db.UsersFunc.SetDefaultReturn(userStore)
+	db.ExternalServicesFunc.SetDefaultReturn(extsvcStore)
+
+	adminCtx := userCtx(1)
+	userCtx := userCtx(2)
+
+	schema, err := graphqlbackend.NewSchema(db, gitserver.NewClient(), nil, graphqlbackend.OptionalResolver{GitHubAppsResolver: NewResolver(logger, db)})
+	require.NoError(t, err)
+
+	graphqlbackend.RunTests(t, []*graphqlbackend.Test{{
+		Schema:  schema,
+		Context: adminCtx,
+		Query: fmt.Sprintf(`
+			query {
+				gitHubApp(id: "%s") {
+					id
+					externalServices(first: 1) {
+						nodes {
+							displayName
+						}
+					}
+				}
+			}`, graphqlID),
+		ExpectedResult: fmt.Sprintf(`{
+			"gitHubApp": {
+				"id": "%s",
+				"externalServices": {
+						"nodes": [{
+							"displayName": "MockExtsvc"
+					}]
+				}
+			}
+		}`, graphqlID),
+	}, {
+		Schema:  schema,
+		Context: userCtx,
+		Query: fmt.Sprintf(`
+			query {
+				gitHubApp(id: "%s") {
+					id
+				}
+			}`, graphqlID),
+		ExpectedResult: `{"gitHubApp": null}`,
+		ExpectedErrors: []*gqlerrors.QueryError{{
+			Message: "must be site admin",
+			Path:    []any{string("gitHubApp")},
 		}},
 	}})
 }