+++ client/cody/CHANGELOG.md
@@ -14,7 +14,12 @@ All notable changes to Sourcegraph Cody will be documented in this file.
 
 ### Fixed
 
+- Error notification display pattern for rate limit [pull/51521](https://github.com/sourcegraph/sourcegraph/pull/51521)
+- Fixes issues with branch switching and file deletions when using the experimental completions feature [pull/51565](https://github.com/sourcegraph/sourcegraph/pull/51565)
+
 ### Changed
+
+- Various improvements to the experimental completions feature
 
 ## [0.0.10]